+++ src/index.ts
@@ -1,6 +1,7 @@
 import { login } from "@utils/loginManager";
 import { loadPlugins } from "@utils/pluginManager";
 import { patchMsgEdit } from "hook/listen";
+import "./hook/message"
 
 // patchMsgEdit();